+++ workalendar/usa/arizona.py
@@ -3,8 +3,10 @@ from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
 from .core import UnitedStates
+from ..registry import iso_register
 
 
+@iso_register('US-AZ')
 class Arizona(UnitedStates):
     """Arizona"""
     martin_luther_king_label = "Dr. Martin Luther King Jr./Civil Rights Day"