+++ workalendar/usa/hawaii.py
@@ -2,10 +2,13 @@
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
+from ..core import FRI
+from ..registry import iso_register
+
 from .core import UnitedStates
-from workalendar.core import FRI
 
 
+@iso_register('US-HI')
 class Hawaii(UnitedStates):
     """Hawaii"""
     include_good_friday = True