+++ workalendar/usa/indiana.py
@@ -4,10 +4,13 @@ from __future__ import (absolute_import, division, print_function,
 
 import warnings
 from datetime import date
-from workalendar.core import MON, TUE, WED, THU, FRI, SAT
+from ..core import MON, TUE, WED, THU, FRI, SAT
+from ..registry import iso_register
+
 from .core import UnitedStates
 
 
+@iso_register('US-IN')
 class Indiana(UnitedStates):
     """Indiana"""
     include_good_friday = True