+++ workalendar/usa/iowa.py
@@ -2,9 +2,11 @@
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
+from ..registry import iso_register
 from .core import UnitedStates
 
 
+@iso_register('US-IA')
 class Iowa(UnitedStates):
     """Iowa"""
     include_thanksgiving_friday = True