+++ workalendar/usa/missouri.py
@@ -3,8 +3,10 @@ from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
 from .core import UnitedStates
+from ..registry import iso_register
 
 
+@iso_register('US-MO')
 class Missouri(UnitedStates):
     """Missouri"""
     FIXED_HOLIDAYS = UnitedStates.FIXED_HOLIDAYS + (